+++ retz-server/src/test/java/io/github/retz/scheduler/RetzSchedulerTest.java
@@ -97,7 +97,7 @@ public class RetzSchedulerTest {
     public void launch() throws InterruptedException, JsonProcessingException, IOException {
         String files[] = {"http://foobar.boom.co.jp/foo.tar.gz"};
         Applications.load(new Application("fooapp", new LinkedList<String>(), new LinkedList<String>(),
-                Arrays.asList(files), Optional.empty(), Optional.empty(), new MesosContainer()));
+                Arrays.asList(files), Optional.empty(), Optional.empty(), "deadbeef", new MesosContainer()));
         Job job = new Job("fooapp", "foocmd", null, new Range(2, 0), new Range(256, 0));
         JobQueue.push(job);
 
@@ -140,7 +140,7 @@ public class RetzSchedulerTest {
     public void notEnough() throws InterruptedException {
         String files[] = {"http://foobar.boom.co.jp/foo.tar.gz"};
         Applications.load(new Application("fooapp", new LinkedList<String>(), new LinkedList<String>(),
-                Arrays.asList(files), Optional.empty(), Optional.empty(), new MesosContainer()));
+                Arrays.asList(files), Optional.empty(), Optional.empty(), "Deadbeef", new MesosContainer()));
         Job job = new Job("fooapp", "foocmd", null, new Range(2, 0), new Range(256, 0));
         JobQueue.push(job);
 
@@ -164,7 +164,7 @@ public class RetzSchedulerTest {
     public void slaveFail() throws InterruptedException, JsonProcessingException, IOException {
         String files[] = {"http://foobar.boom.co.jp/foo.tar.gz"};
         Applications.load(new Application("fooapp", new LinkedList<String>(), new LinkedList<String>(),
-                Arrays.asList(files), Optional.empty(), Optional.empty(), new MesosContainer()));
+                Arrays.asList(files), Optional.empty(), Optional.empty(), "Deadbeef", new MesosContainer()));
         Job job = new Job("fooapp", "foocmd", null, new Range(2, 0), new Range(256, 0));
         JobQueue.push(job);